requirements.txt
certifi==2020.12.5
owoify-py==1.0.31
python-dotenv==0.15.0
slack-bolt==1.3.1
slack-sdk==3.3.1
